handle existing user entry when creating account
till now the only way to create an entry in the user table was via
create-account. but we are going to import email addresses of people
who signup to our mailing list, and store them also in the same
`fastn_user` table.

we have to make sure that if an email was already imported and user is
creating an account, we do not create a new account, and instead put the
information in the same account.

it is possible that email was double opt-in, verified, and someone else
enters the same email address, and we still have to do confirm account,
the fact is pre-verified can not be used during account creation
process.

in future we will also allow another url, which can be constructed when
we are sending emails, which is pre-verified, in that case when the user
sees the account creation form they will only be able to edit username,
name etc (name can be prepopulated if user entered it during email
signup flow), and no confirmation mail will be sent.