handle diesel::NotFound in create-account
the query:

```sql
    SELECT
        id, identity
    FROM fastn_user
    WHERE
        EXISTS (
            SELECT 1
            FROM json_each(data -> 'email' -> 'emails' )
            WHERE value = $1
        )
    LIMIT 1
```

may fail and diesel::NotFound will be returned if `.optional()` is not
used.

The query is wrong anyways. As a bad actor, I can just register with
random emails and don't verify myself. The genuine owner of this
email won't be able to create an account because someone already
registered for them but did not verify!