merge create_verified_account into create_account
there was not much difference between these two and it made the frontend
code complex having two different handlers for creating users of
different type (subscribers and new users). Now only create_account
exists which handles both the scenarios.

The query param ?code is always expected with requests. This is not a
problem as the frontend will send a garbage ?code anyway.