Fix: Use Non Strict Secure cookies to allow cross-site flows
The auth cookie is not attached to a request in cross-origin contexts.
This is bad as we can't link to a website that is relying on auth data
from another website.

Auth cookies are in general Non Strict but Secure.

This is provided as a direct change on the tip of 0.1.2 version so that
websites currently on 0.1.2 can upgrade to this. A major release with
`main` will be made in future with a proper release plan.