Fix: Confirm account link
The link did not have the name of the wasm module (email_auth_provider)
so /confirm-email/ route handler was never hit. This is fixed now.

The change assumes that the name of the wasm module will always be
email_auth_provider.wasm and path
lets-auth.fifthtry.site/email_auth_provider.wasm must exist.