Feat: /forgot-password/ app url
A forgot password page is added to the lets-auth. Anyone can request a
password reset link in their email using this page. It supports both
username and email.

After link is generated and sent to user's email, the user is redirected
to /forgot-password-success/

The backend handlers don't rely on query params for route links. We use
app-url to assume that the password reset will happen on
/reset-password/ page (relative to wherever lets-auth is mounted)