Fix: Remove unused config keys from /config
The keys are not used anymore by the backend. The values that are
removed use the `ftd.app-url` function which errors when used with
`ftd.json`. As we don't need this, it's simply removed from /config for
now. `fastn` should fix this bug and support ftd.app-url in ftd.json.