Feat: `lets_auth::folder::has_perms_on_folder` fn
This function checks if the user has a permission on a folder. The user
must belong to the folder.

Some new types are also introduced to support this function. Some may
even clash with existing `Folder` type (calling it `FolderInput` for now
to avoid this).

`denormalized_folders` function impl is commented out. This is not
needed right now and breaks compilation.